+++ packages/definitions-parser/test/git.test.ts
@@ -1,8 +1,8 @@
 import * as util from "util";
 import * as pacote from "pacote";
 import { createTypingsVersionRaw, testo } from "./utils";
-import { GitDiff, getNotNeededPackages, checkNotNeededPackage } from "../src/git";
-import { NotNeededPackage, AllPackages } from "../src/packages";
+import { GitDiff, getNotNeededPackages, checkNotNeededPackage, gitChanges } from "../src/git";
+import { NotNeededPackage, AllPackages, PackageId } from "../src/packages";
 
 const typesData = {
   jquery: createTypingsVersionRaw("jquery", {}, {}),
@@ -21,43 +21,103 @@ const deleteJestDiffs: GitDiff[] = [
   { status: "D", file: "types/jest/index.d.ts" },
   { status: "D", file: "types/jest/jest-tests.d.ts" },
 ];
+const moveRdfJSDiffs: GitDiff[] = [
+  { status: "D", file: "types/rdf-ext/ClownfaceFactory.d.ts" },
+  { status: "A", file: "types/rdf-ext/FetchFactory.d.ts" },
+  { status: "A", file: "types/rdf-ext/FormatsFactory.d.ts" },
+  { status: "M", file: "types/rdf-ext/index.d.ts" },
+  { status: "M", file: "types/rdf-ext/package.json" },
+  { status: "M", file: "types/rdf-ext/rdf-ext-tests.ts" },
+  { status: "D", file: "types/rdfjs__environment/DataFactory.d.ts" },
+  { status: "D", file: "types/rdfjs__environment/DatasetFactory.d.ts" },
+  { status: "D", file: "types/rdfjs__environment/NamespaceFactory.d.ts" },
+  { status: "D", file: "types/rdfjs__environment/TermMapSetFactory.d.ts" },
+  { status: "M", file: "types/rdfjs__environment/package.json" },
+  { status: "M", file: "types/rdfjs__environment/rdfjs__environment-tests.ts" },
+  { status: "R", source: "types/rdfjs__formats-common/.npmignore", file: "types/rdfjs__formats/.npmignore" },
+  { status: "R", source: "types/rdfjs__environment/FormatsFactory.d.ts", file: "types/rdfjs__formats/Factory.d.ts" },
+  { status: "R", source: "types/rdfjs__formats-common/index.d.ts", file: "types/rdfjs__formats/index.d.ts" },
+  { status: "R", source: "types/rdfjs__environment/lib/Formats.d.ts", file: "types/rdfjs__formats/lib/Formats.d.ts" },
+  { status: "R", source: "types/rdfjs__formats-common/package.json", file: "types/rdfjs__formats/package.json" },
+  { status: "A", file: "types/rdfjs__formats/pretty.d.ts" },
+  {
+    status: "R",
+    source: "types/rdfjs__formats-common/rdfjs__formats-common-tests.ts",
+    file: "types/rdfjs__formats/rdfjs__formats-tests.ts",
+  },
+  { status: "A", file: "types/rdfjs__formats/tsconfig.json" },
+  { status: "M", file: "types/rdfjs__serializer-jsonld-ext/index.d.ts" },
+  { status: "M", file: "types/rdfjs__serializer-jsonld-ext/package.json" },
+  { status: "M", file: "types/rdfjs__serializer-jsonld-ext/rdfjs__serializer-jsonld-ext-tests.ts" },
+  { status: "A", file: "types/rdfjs__serializer-turtle/.npmignore" },
+  { status: "A", file: "types/rdfjs__serializer-turtle/index.d.ts" },
+  { status: "A", file: "types/rdfjs__serializer-turtle/package.json" },
+  { status: "A", file: "types/rdfjs__serializer-turtle/rdfjs__serializer-turtle-tests.ts" },
+  {
+    status: "R",
+    source: "types/rdfjs__formats-common/tsconfig.json",
+    file: "types/rdfjs__serializer-turtle/tsconfig.json",
+  },
+];
+function getDeletions(diffs: GitDiff[]): PackageId[] {
+  const changes = gitChanges(diffs);
+  expect(changes).not.toHaveProperty("error");
+  const { deletions } = changes as { deletions: PackageId[]; additions: PackageId[] };
+  return deletions;
+}
 
 testo({
   async ok() {
-    expect(await getNotNeededPackages(allPackages, deleteJestDiffs)).toEqual(jestNotNeeded);
+    expect(await getNotNeededPackages(allPackages, getDeletions(deleteJestDiffs))).toEqual(jestNotNeeded);
+  },
+  async gitMovesConvertedToAddsAndDeletes() {
+    expect(gitChanges(moveRdfJSDiffs)).toEqual({
+      additions: [
+        { typesDirectoryName: "rdf-ext", version: "*" },
+        { typesDirectoryName: "rdfjs__formats", version: "*" },
+        { typesDirectoryName: "rdfjs__serializer-turtle", version: "*" },
+      ],
+      deletions: [
+        { typesDirectoryName: "rdf-ext", version: "*" },
+        { typesDirectoryName: "rdfjs__environment", version: "*" },
+        { typesDirectoryName: "rdfjs__formats-common", version: "*" },
+      ],
+    });
   },
   async forgotToDeleteFiles() {
     expect(
       await getNotNeededPackages(
         AllPackages.fromTestData({ jest: createTypingsVersionRaw("jest", {}, {}) }, jestNotNeeded),
-        deleteJestDiffs,
+        getDeletions(deleteJestDiffs),
       ),
     ).toEqual({ errors: ["Please delete all files in jest when adding it to notNeededPackages.json."] });
   },
   async tooManyDeletes() {
-    expect(await getNotNeededPackages(allPackages, [{ status: "D", file: "types/oops/oops.txt" }])).toEqual([]);
+    expect(
+      await getNotNeededPackages(allPackages, getDeletions([{ status: "D", file: "types/oops/oops.txt" }])),
+    ).toEqual([]);
   },
   async deleteInOtherPackage() {
     expect(
-      await getNotNeededPackages(allPackages, [
-        ...deleteJestDiffs,
-        { status: "D", file: "types/most-recent/extra-tests.ts" },
-      ]),
+      await getNotNeededPackages(
+        allPackages,
+        getDeletions([...deleteJestDiffs, { status: "D", file: "types/most-recent/extra-tests.ts" }]),
+      ),
     ).toEqual(jestNotNeeded);
   },
   async extraneousFile() {
     expect(
-      await getNotNeededPackages(allPackages, [
-        ...deleteJestDiffs,
-        { status: "A", file: "types/oops/oooooooooooops.txt" },
-      ]),
+      await getNotNeededPackages(
+        allPackages,
+        getDeletions([...deleteJestDiffs, { status: "A", file: "types/oops/oooooooooooops.txt" }]),
+      ),
     ).toEqual(jestNotNeeded);
   },
   async scoped() {
     expect(
       await getNotNeededPackages(
         AllPackages.fromTestData(typesData, [new NotNeededPackage("ember__object", "@ember/object", "1.0.0")]),
-        [{ status: "D", file: "types/ember__object/index.d.ts" }],
+        getDeletions([{ status: "D", file: "types/ember__object/index.d.ts" }]),
       ),
     ).toEqual([new NotNeededPackage("ember__object", "@ember/object", "1.0.0")]);
   },