+++ packages/eslint-plugin/test/fixtures/types/expect-diagnostics-editor/index.d.ts
@@ -1,0 +1,4 @@
+export const foo = 1234;
+
+// No diagnostic; we did not set versionsToTest and so are in an editor.
+export const badSet: Array<string, number, number, number>;