+++ packages/eslint-plugin/test/fixtures/types/no-type-only-packages-reexport/index.d.ts
@@ -1,0 +1,3 @@
+import { SyntaxKind } from "typescript";
+
+export { SyntaxKind };