+++ packages/eslint-plugin/test/fixtures/types/no-type-only-packages/index.d.ts
@@ -1,0 +1,11 @@
+export interface Blah {
+    foo: string;
+}
+
+export namespace Something {
+    export interface Blah2 {
+        foo: string;
+    }
+}
+
+export type NotAValue = Something.Blah2;