+++ packages/dts-gen/CHANGELOG.md
@@ -1,5 +1,17 @@
 # dts-gen
 
+## 0.9.0
+
+### Minor Changes
+
+- 2d7a5d3: Require Node 18+
+
+## 0.8.2
+
+### Patch Changes
+
+- 2c3e5de: Update dependencies
+
 ## 0.8.1
 
 ### Patch Changes