+++ packages/header-parser/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @definitelytyped/header-parser
 
+## 0.2.0
+
+### Minor Changes
+
+- 2d7a5d3: Require Node 18+
+
+### Patch Changes
+
+- Updated dependencies [2d7a5d3]
+  - @definitelytyped/typescript-versions@0.1.0
+  - @definitelytyped/utils@0.1.0
+
+## 0.1.2
+
+### Patch Changes
+
+- 2c3e5de: Update dependencies
+- Updated dependencies [2c3e5de]
+  - @definitelytyped/utils@0.0.192
+
 ## 0.1.1
 
 ### Patch Changes